python/pyproject.toml
[project]
name = "creda"
version = "2.0.0"
description = "CREDA — AI-powered multilingual financial coach for India"
requires-python = ">=3.12,<3.14"

dependencies = [
    # ── Django Frontend ──
    "django>=5.1",
    "django-htmx>=1.21",
    "whitenoise>=6.8",
    "httpx>=0.27",
    "pyjwt>=2.9",
    # ── FastAPI Backend ──
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis>=5.2",
    # ── LLM / Agents ──
    "langchain-groq>=0.2",
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "groq>=0.15",
    "chromadb>=1.0",
    "sentence-transformers>=2.7",
    # ── Voice ──
    "faster-whisper>=1.1",
    "edge-tts>=7.0",
    "pydub>=0.25",
    # ── Finance ──
    "casparser>=0.8",
    "pyxirr>=0.10",
    "yfinance>=0.2",
    "numpy>=1.26",
    "numpy-financial>=1.0",
    # ── Utilities ──
    "python-dotenv>=1.0",
    "python-multipart>=0.0.12",
    "slowapi>=0.1",
    "apscheduler>=3.10",
    "langdetect>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",             # for TestClient
    "ruff>=0.8",
]
whatsapp = [
    "twilio>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
